+++ src/adapter/stream.rs
@@ -9,6 +9,7 @@ pub struct LoggingStream<S> {
     inner: S,
     collected_content: Arc<Mutex<String>>,
     db_pool: Arc<DatabasePool>,
+    service_id: Option<String>,
     provider_id: i64,
     provider_name: String,
     model: String,
@@ -23,6 +24,7 @@ impl<S> LoggingStream<S> {
         inner: S,
         collected_content: Arc<Mutex<String>>,
         db_pool: Arc<DatabasePool>,
+        service_id: Option<String>,
         provider_id: i64,
         provider_name: String,
         model: String,
@@ -33,6 +35,7 @@ impl<S> LoggingStream<S> {
             inner,
             collected_content,
             db_pool,
+            service_id,
             provider_id,
             provider_name,
             model,
@@ -56,6 +59,7 @@ impl<S: Stream + Unpin> Stream for LoggingStream<S> {
                     .unwrap_or(None);
 
                 let log = NewRequestLog {
+                    service_id: self.service_id.clone(),
                     provider_id: Some(self.provider_id),
                     provider_name: self.provider_name.clone(),
                     model: self.model.clone(),